Refactor: BeanPayDetail 생성 로직 리펙토링
